backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
langchain==0.2.7
langchain-openai==0.1.15
langchain-community==0.2.7
faiss-cpu==1.8.0.post1
PyPDF2==3.0.1
pymongo==4.7.3
cryptography==42.0.8
python-dotenv==1.0.1

